requirements.txt
amazon-dash==0.2.2
certifi==2017.11.5
chardet==3.0.4
idna==2.6
PyYAML==3.12
requests==2.18.4
scapy-python3==0.21
urllib3==1.22
